mtd: ubi: add a ubi forced detach ioctl
When running ubi rootfs upgrade on nand based OpenWrt device after
pivot_root init process is still "hooked" on the old file system.
Thus, the old file system can not be umounted. If the filesystem is
mounted it can not be upgraded with using for example ubiupdatevol
or removed with ubirmvol. Forcing umount would allow to run the
before mentioned commands.

Change-Id: I431c5957aea1b2f1f7cc7645ff7a1ae5b22e6d35
Signed-off-by: Ram Chandra Jangir <quic_rjangir@quicinc.com>