net: nat44: Fix for Linux NAPT limitaion.
When multiple iptables rules are configured with different port-ranges,
kernel code doesn't fall through to next rule in the chain, even if all
ports available in the current rule are exhausted, so the kernel makes
the flow to share the same snat 5-tuple with other flows.

This fix returns XT_CONTINUE when the tuple returned by get_unique_tuple
is already used. Hence, it will move to next rule available in the chain.

NF_NAT_TRY_NEXT_RULE in netfilter/Kconfig should be set to 'y' prior to
using this fix.

Change-Id: I414887198ecf367c1be97f6f9495b4eb8d36b689
Signed-off-by: Apoorv Gupta <apoogupt@codeaurora.org>

net: nat44: Fix for NAPT limitation for ported protocols only.

When multiple iptables nat rules are configured with different port-ranges,
kernel code doesn't fall through to next rule in the chain, even if all
ports available in the current rule are exhausted, so the kernel makes
the flow to share the same 5-tuple with other flows.

This fix returns XT_CONTINUE for ported protocols only when the tuple
returned by get_unique_tuple is already used. Hence, it will move to next
rule available in the chain.

CONFIG_NF_NAT_TRY_NEXT_RULE in netfilter/Kconfig should be set to 'y'
prior to using this fix.

Change-Id: I6d9e79541359266ed9ec9d5420b7346801d528e1
Signed-off-by: Apoorv Gupta <apoogupt@codeaurora.org>