OpenWrt: 191-rtc-rs5c372-let_the_alarm_to_be_used_as_wakeup_source.patch
rtc: rs5c372: let the alarm to be used as wakeup source

Currently there is no use for the interrupts on the rs5c372 RTC and the
wakealarm isn't enabled. There are some devices like NASes which use this
RTC to wake up from the power off state when the INTR pin is activated by
the alarm clock.

Enable the alarm and let to be used as a wakeup source.

Tested on a Buffalo LS421DE NAS.

Signed-off-by: Daniel González Cabanelas <dgcbueu@gmail.com>

Change-Id: If4bae9db7a2c1b4a32fd54dca23f774585c30572
Signed-off-by: Ram Chandra Jangir <quic_rjangir@quicinc.com>