OpenWrt: 682-of_net-add-mac-address-increment-support.patch
of_net: add mac-address-increment support

Lots of embedded devices use the mac-address of other interface
extracted from nvmem cells and increments it by one or two. Add two
bindings to integrate this and directly use the right mac-address for
the interface. Some example are some routers that use the gmac
mac-address stored in the art partition and increments it by one for the
wifi. mac-address-increment-byte bindings is used to tell what byte of
the mac-address has to be increased (if not defined the last byte is
increased) and mac-address-increment tells how much the byte decided
early has to be increased.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>

Change-Id: Id2abf7f92f2590d06c2a80cb4a2fdf9eecb6788b
Signed-off-by: Ram Chandra Jangir <quic_rjangir@quicinc.com>