OpenWrt: 834-ledtrig-libata.patch
libata: add ledtrig support

This adds a LED trigger for each ATA port indicating disk activity.

As this is needed only on specific platforms (NAS SoCs and such),
these platforms should define ARCH_WANTS_LIBATA_LEDS if there
are boards with LED(s) intended to indicate ATA disk activity and
need the OS to take care of that.
In that way, if not selected, LED trigger support not will be
included in libata-core and both, codepaths and structures remain
untouched.

Signed-off-by: Daniel Golle <daniel@makrotopia.org>

Change-Id: I91046656b4c57a66e1c5f7fe792154a7e8bfe5bc
Signed-off-by: Ram Chandra Jangir <quic_rjangir@quicinc.com>