OpenWrt: 205-kconfig-abort-configuration-on-unset-symbol.patch
kconfig: abort configuration on unset symbol

When a target configuration has unset Kconfig symbols, the build will
fail when OpenWrt is compiled with V=s and stdin is connected to a tty.

In case OpenWrt is compiled without either of these preconditions, the
build will succeed with the symbols in question being unset.

Modify the kernel configuration in a way it fails on unset symbols
regardless of the aforementioned preconditions.

Signed-off-by: David Bauer <mail@david-bauer.net>

Change-Id: If3f6f7c3a2af97b4fc32d5f397232e5aa19f4eca
Signed-off-by: Ram Chandra Jangir <quic_rjangir@quicinc.com>