OpenWrt: 421-drivers-mtd-parsers-add-nvmem-support-to-cmdlinepart.patch
drivers: mtd: parsers: add nvmem support to cmdlinepart

Assuming cmdlinepart is only one level deep partition scheme and that
static partition are also defined in DTS, we can assign an of_node for
partition declared from bootargs. cmdlinepart have priority than
fiexed-partition parser so in this specific case the parser doesn't
assign an of_node. Fix this by searching a defined of_node using a
similar fixed_partition parser and if a partition is found with the same
label, check that it has the same offset and size and return the DT
of_node to correctly use NVMEM cells.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>

Change-Id: Ie1646192423098cbdc5a090f9b4574eac8721d3c
Signed-off-by: Ram Chandra Jangir <quic_rjangir@quicinc.com>