remoteproc: qcom: Add NOTIFY_FATAL event type to SSR subdevice
Currently the SSR subdevice notifies the client driver on crash of the
rproc from the recovery workqueue using the BEFORE_SHUTDOWN event.
However the client driver might be interested to know that the device
has crashed as soon as possible atomically.

Add NOTIFY_FATAL event to SSR subdevice to atomically notify rproc has
crashed to the client driver.

Change-Id: I5ba63bff6c93cf41c23945d9691e39f402398616
Signed-off-by: Vignesh Viswanathan <quic_viswanat@quicinc.com>