ppp: PPPoE acceleration support.
pppoe: Add return value to pppoe addressing get function

If the addressing doesn't have a netdevice, this should be
handled as a failure, so that the caller considers that
the get function failed.

Change-Id: Ia9a6b0e0f036a3434519d9f2194763486ca04583
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>

Added some new APIs to the PPP/PPPoE kernel modules
for using from the hardware acceleration connection managers.

Change-Id: I2c16c6d6ccba8ffa14aec077c8dad1681535ae0b
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>

net: ppp: rx/tx error and dropped pkt stats support

ppp_update_stats functions accepts rx/tx errors and dropped
pkt stats args.

Change-Id: Iba2f6ea2114d8a4678254332fec0ef7bc35bed2c
Signed-off-by: ratheesh kannoth <rkannoth@codeaurora.org>
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>

ppp: Update the last_recv and last_xmit times.

These need to be updated for accelerated connections, so that
on demand mode will recognize the active traffic.

Change-Id: I3c0ee4e8f4c3bc4c7ce221e6109bfd82046d11b4
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>

ppp: Add channel connect/disconnect notifier.

When a channel is connected or disconnected to a PPP
unit, a notification will be sent to the kernel subsystems
which are registered to this notifier. Notifier sends
the event type (connect or disconnect) along with the
PPP netdevice pointer.

Change-Id: I0d592084744c86e1c114ac66da68b214ddff5b5a
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>
Signed-off-by: Murat Sezgin <quic_msezgin@quicinc.com>