clk: qcom: apss-ipq6018: add support for the switch between the safe source
Stromer Plus PLL found on IPQ53xx doesn't support the dynamic frequency
scaling. To achieve the same, we need to park the APPS PLL source to GPLL0,
re configure the PLL and then switch the source to APSS_PLL_EARLY.

To support this, register a clock notifier to get the PRE_RATE and POST_RATE
notification. Change the APSS PLL source to GPLL0 when PRE_RATE notification
is received, then configure the PLL and then change back the source to
APSS_PLL_EARLY.

Change-Id: I2fc8d169ac5b235431680d150992139cae5f214d
Signed-off-by: Kathiravan T <quic_kathirav@quicinc.com>