net: qrtr: ns: Change nodes radix tree to xarray
There is a use after free scenario while iterating through the servers
radix tree despite the ns being a single threaded process. This can
happen when the radix tree APIs are not synchronized with the
rcu_read_lock() APIs.

Convert the radix tree for nodes to xarray to take advantage of the
built in rcu lock usage provided by xarray.

Signed-off-by: Chris Lew <clew@codeaurora.org>
Signed-off-by: Saahil Tomar <quic_saahtoma@quicinc.com>
Change-Id: Ia443eae33a5ae6b8857e2416d42dcd3a39ce4bab