soc: introduce config QTI_APSS_ACC
cpu_ops and cpu_pwr_ctl is required only for the stand alone boot
up of secondary cores. With the full boot chain, this driver is not
required since the PSCI is used. So put this driver under a config
and enable it only when required.

This change also fixes the compilation issues when the QTI_APSS_ACC
is enabled.

Reference Commit: I2d0c85e10c

Change-Id: I6faa4a979174c27cde1b39bd302dd8b7189bdb9c
Signed-off-by: Timple Raj M <quic_timple@quicinc.com>