pcie: qcom: ipq9574: Forcing PCIE2/3 to single lane
For lane switching, as num-lanes property is not
supported in Alder. This patch enables the support
for switching PCIE2/3 to single lane. The function
is added in PCI_FINAL_FIXUP and can be enabled from
the DTS node by defining "force_to_single_lane" property.

NOTE: Do not change num-lanes to 1 for PCIE2/3.
Instead, "force_to_single_lane" property can be defined
in DTS for PCIE2/3 to switch the corresponding node to
single lane.

This is a snapshot of the commit
42f4e796a2bbae8c3d10962e8d36fb0526ecada2
from win.linuxopenwrt.2.1 branch and
oss/kernel/linux-ipq-5.4 project.

Change-Id: I8b7247c20dff1a926a597584da3ad34d11a3309f
Signed-off-by: Hariharan K <quic_harihk@quicinc.com>