drivers: mtd: qcom_nandc: Add switch layout support
In ipq9574 platform, PBL doesn't have 4K nand support.

In order to support 4K QSPI nand in such a platform,
we have introduced this layout switch mechanism. In this,
the SBL will be in 2K and all other partitions would be
in 4K. Therefore, this change is required to make sure
during sysupgrade the layout is switched as needed
accordingly for SBL and for other partitions.

Change-Id: Ied559e9881fdc04ae0fdaaddf6fcaff0712d0ecd
Signed-off-by: Sridharan S N <quic_sridsn@quicinc.com>